requirements.txt
langchain
langchain_groq
streamlit_option_menu
pandas
numpy
streamlit
scikit-learn
seaborn
plotly
python-dotenv
streamlit-aggrid
